Rename Fw/SMTest to Fw/Sm
