+++ Fw/CMakeLists.txt
@@ -14,7 +14,7 @@ add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/Tlm/")
 # Framework subdirectories
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/Cfg/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/Comp/")
-add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/SMSignal/")
+add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/Sm/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/FilePacket/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/Obj/")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/Port/")
