--- Fw/SMSignal/docs/.gitignore
@@ -1,1 +1,0 @@
-*.html
