+++ Fw/Sm/SMSignalBuffer.hpp
